refactor(core): split peers manager actor in different files
